+++ base/src/bit.rs
@@ -3,6 +3,7 @@
 //! Most traits are only implemented for unsigned integers yet.
 
 /// Common bit operations for integers
+// TODO(v0.3): add doc tests
 pub trait BitTest {
     /// Get the minimum required number of bits to represent this integer
     fn bit_len(&self) -> usize;
@@ -12,9 +13,11 @@ pub trait BitTest {
 
     /// Get the number of trailing zeros in the integer
     fn trailing_zeros(&self) -> Option<usize>;
+
+    // TODO(v0.3): add trailing_ones
 }
 
-/// Next power of two.
+/// Functions related to the power of two.
 ///
 /// # Examples
 /// ```
@@ -25,7 +28,9 @@ pub trait BitTest {
 /// assert_eq!(n.next_power_of_two(), 8);
 /// ```
 pub trait PowerOfTwo {
+    /// Test if self is a power of two (`2^k`)
     fn is_power_of_two(&self) -> bool;
+    /// Get the smallest power of two greater than or equal to self.
     fn next_power_of_two(self) -> Self;
 }