+++ float/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.1
+
+- Implement `core::iter::{Sum, Product}` for `FBig`
+- Implement `powf`, `sqrt` for `FBig`
+
 ## 0.2.0 (Initial release)
 
 - Support basic arithmetic operations (`add`/`sub`/`mul`/`div`/`exp`/`ln`) and base conversion.
@@ -8,10 +13,11 @@
 
 ## Roadmap to next version
 - Support generating base math constants (E, Pi, SQRT2, etc.)
+- Support operations with inf
 - Implement Random generator
 - Implement Serde serialization
-- Implememt sqrt, cbrt, nth_root
-- Implement powf, log
+- Implememt cbrt, nth_root
+- Implement log
 - Create operations benchmark
 - Benchmark against crates: rug, twofloat, num-bigfloat, rust_decimal, bigdecimal, scientific
 - Implement more formatting traits