+++ integer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.2.1
+
+### Add
+
+- Add `sqrt`, `sqrt_rem`, `nth_root` for `UBig` and `IBig`
+- Implement `core::iter::{Sum, Product}` for `UBig` and `IBig`
+- Implement `num_traits::{Euclid, ToPrimitive, FromPrimitive}` for `UBig` and `IBig`
+- Implement `num_integer::{Integer, Roots}` for `UBig` and `IBig`
+- Implement `num_order::{NumHash, NumOrd}` for `UBig` and `IBig`
+- Implement `zeroize::Zeroize` for `UBig`, `IBig` and internal types
+- `serde` se/derialization now supports the `is_human_readable()` flag
+
 ## 0.2.0
 
 ### Add